Changing the format of the Excel sheet
All the prizes are now in a single sheet, refer to the template to see how to build one.

Changing the format of the PowerPoint too so it looks cleaner, now the number of people per slide is not continuous anymore, so having a slide with a different number of winners requires changing the slideLayouts array in the code.